\name{print.summary.kingui0}
\alias{print.summary.kingui0}
\title{Print method for \code{\link{summary.kingui}}}
\usage{
  print.summary.kingui0(x,
    digits = max(3, getOption("digits") - 3),
    detailed = FALSE, ...)
}
\arguments{
  \item{x}{An object of class \code{summary.kingui}}

  \item{digits}{How many digits should be printed after the
  decimal point.}

  \item{detailed}{Ignore for now.}

  \item{...}{Other parameters to be passed into
  \code{\link{summary.kingui}}}
}
\value{
  \code{NULL}
}
\description{
  ##' Expanded from \code{\link{print.summary.modFit}} and
  \code{\link{print.summary.mkinfit}}
}
\author{
  Zhenglei Gao
}
\keyword{internal}

